Add files via upload
The main task of this project is to analyze the sales data of Baja for the trading period 01/01/21 to 31/03/21, from nine separate Baja stores in Scotland. The analysis aims to demonstrate the potential of business analytics in understanding the business and facilitating informed decision-making.

The following are the specific objectives of this analysis:

Data Inspection and Cleaning: The first step would involve examining the provided data and performing necessary cleaning operations. Remarks about the overall validity of the data will be discussed.

Development of an Analytics Dashboard: Development of an Excel-based dashboard that summarizes the key findings from the dataset and allows for further detailed investigations.

Risk Analysis using Monte Carlo simulations: Perform a risk assessment of profitability levels in the Aberdeen Homeware department with Monte Carlo simulations. The analysis will only consider non-promotion sales and will assume gross profit margins between 60% to 70%.

Transportation Cost Optimization: Use optimization processes to reduce the transportation costs of food from the northern distribution centers to the stores, while ensuring the centers can deliver at least the equivalent of the maximum non-promotion day sales to each store.

Critical Evaluation of Data Impact: Critically evaluate and explain the impact of the data on various aspects of business and the benefits derived from it.

Overall, the aim is to leverage analytics to draw actionable insights from sales data that can support better decision-making for the business.